refactor(atc): point the heartbeat timer at the shared unit-program module
#609 and this branch fixed the same defect in two places and grew two
copies of the machinery that does it: choosing the binary, wrapping it
in /bin/sh -c, and parsing it back out of an installed unit to check it
resolves. Now that #609 has landed, collapse the timer onto the shared
module rather than leaving the copies to drift.

Deleted from timer.rs and taken from crate::fleet::unit_program:
SHELL, ainb_bin_from, unit_path_env, ProgramHealth, unit_program_health,
unit_argv, program_from_argv, is_shell, unit_search_path,
program_resolves_in, plist_program_arguments, plist_string_after_key,
systemd_exec_start_argv, systemd_environment_path, xml_unescape,
shell_split, shell_quote.

Behaviour is unchanged - the shared copies came from #609 - so this is
a move, not a rewrite. ProgramHealth is re-exported so the return type
of installed_program_health stays reachable as timer::ProgramHealth and
cli/fleet/atc.rs needs no edit.

Timer tests that only exercised the moved helpers are dropped; the
equivalents live in unit_program. Tests that exercise the timer's own
unit generation stay, with the two that reached into now-private
parsers rewritten onto unit_program_health. The one junk-input case
that was only covered in timer.rs is folded into the shared test.

timer.rs 872 -> 517 lines. 35 tests pass across the three modules.

Refs #608
